requirements.txt
pdfminer.six==20191125
PyPDF2==3.0.1
thefuzz==0.20.0
Unidecode==1.3.8
